src/requirements.txt
Flask==1.0.2
numpy==1.20.2
Pillow==2.2.2